app/services/image_to_3d_service.py
import os
import torch
import numpy as np
from PIL import Image
import logging
import uuid
from typing import Dict, Any, Optional, Tuple

class ImageTo3DService:
    """Service for converting images to 3D models using Hunyuan"""
    
    def __init__(self, config: Dict[str, Any]):
        self.logger = logging.getLogger("image_to_3d_service")
        self.config = config
        self.model_path = config.get("HUNYUAN_MODEL_PATH", "app/models/hunyuan_model")
        self.output_dir = config.get("MODELS_DIR", "data/models")
        
        # Device configuration
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        self.logger.info(f"Using device: {self.device}")
        
        # Load Hunyuan model
        self._load_model()
    
    def _load_model(self):
        """Load the Hunyuan model"""
        try:
            self.logger.info("Loading Hunyuan model...")
            
            # This is a placeholder for actual Hunyuan model loading
            # Replace with actual model loading code based on Hunyuan documentation
            
            # Example of loading a model with torch:
            # self.model = torch.load(os.path.join(self.model_path, "model.pth"), map_location=self.device)
            # self.model.eval()
            
            # For now, we'll just set a flag
            self.model_loaded = True
            self.logger.info("Hunyuan model loaded successfully")
            
        except Exception as e:
            self.logger.error(f"Error loading Hunyuan model: {str(e)}")
            self.model_loaded = False
            raise
    
    def process_image(self, image_path: str, params: Dict[str, Any]) -> Tuple[bool, Dict[str, Any]]:
        """
        Process an image and generate a 3D model
        
        Args:
            image_path: Path to the input image
            params: Processing parameters (resolution, detail level, etc.)
            
        Returns:
            Tuple of (success, result_data)
        """
        try:
            if not self.model_loaded:
                return False, {"error": "Model not loaded"}
            
            self.logger.info(f"Processing image: {image_path}")
            
            # Load and preprocess the image
            image = Image.open(image_path).convert("RGB")
            
            # Extract parameters with defaults
            resolution = int(params.get("resolution", 128))
            detail_level = float(params.get("detail_level", 0.5))
            smoothing = float(params.get("smoothing", 0.3))
            
            # Preprocess image
            processed_image = self._preprocess_image(image, resolution)
            
            # In a real implementation, you would:
            # 1. Run the image through the Hunyuan model
            # 2. Process the output into a 3D mesh
            # 3. Apply post-processing (smoothing, decimation)
            # 4. Save in appropriate formats
            
            # For this example, we'll create a placeholder result
            model_id = f"img3d_{uuid.uuid4().hex[:8]}"
            output_path = os.path.join(self.output_dir, f"{model_id}")
            os.makedirs(output_path, exist_ok=True)
            
            # Save output files (placeholder)
            obj_path = os.path.join(output_path, f"{model_id}.obj")
            stl_path = os.path.join(output_path, f"{model_id}.stl")
            
            # Create a simple cube mesh as placeholder
            # In real implementation, this would be the output from Hunyuan
            with open(obj_path, 'w') as f:
                f.write("""
# Simple cube OBJ file (placeholder)
v 0 0 0
v 1 0 0
v 1 1 0
v 0 1 0
v 0 0 1
v 1 0 1
v 1 1 1
v 0 1 1
f 1 2 3
f 1 3 4
f 5 6 7
f 5 7 8
f 1 2 6
f 1 6 5
f 2 3 7
f 2 7 6
f 3 4 8
f 3 8 7
f 4 1 5
f 4 5 8
""")

            # Create STL file (simple placeholder)
            with open(stl_path, 'w') as f:
                f.write("solid cube\n")
                f.write("  facet normal 0 0 0\n")
                f.write("    outer loop\n")
                f.write("      vertex 0 0 0\n")
                f.write("      vertex 1 0 0\n")
                f.write("      vertex 1 1 0\n")
                f.write("    endloop\n")
                f.write("  endfacet\n")
                # Add more facets for a full cube...
                f.write("endsolid cube\n")
                
            # Create a preview image
            preview_path = os.path.join(output_path, "preview.png")
            image.resize((256, 256)).save(preview_path)
            
            self.logger.info(f"3D model generated: {model_id}")
            
            return True, {
                "model_id": model_id,
                "obj_path": obj_path,
                "stl_path": stl_path,
                "preview_path": preview_path,
                "parameters": params,
                "processing_time": 2.5  # Placeholder
            }
            
        except Exception as e:
            self.logger.error(f"Error processing image: {str(e)}")
            return False, {"error": str(e)}
    
    def _preprocess_image(self, image, resolution):
        """Preprocess image for the model"""
        # Resize to target resolution
        image = image.resize((resolution, resolution), Image.LANCZOS)
        
        # Convert to numpy array and normalize
        image_array = np.array(image).astype(np.float32) / 255.0
        
        # If using PyTorch, convert to tensor
        # image_tensor = torch.from_numpy(image_array).permute(2, 0, 1).unsqueeze(0)
        # image_tensor = image_tensor.to(self.device)
        
        return image_array
    
    def generate_preview(self, model_id: str) -> Optional[str]:
        """Generate a preview image of the 3D model"""
        try:
            # In a real implementation, you would render the 3D model
            # and save it as an image
            
            # For now, we'll just return a placeholder path
            preview_path = os.path.join(self.output_dir, f"{model_id}", "preview.png")
            return preview_path
            
        except Exception as e:
            self.logger.error(f"Error generating preview: {str(e)}")
            return None
