requirements.txt
# Existing requirements
flask>=2.0.1
pyjwt>=2.1.0
razorpay>=1.3.0
python-dotenv>=0.19.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=9.0.0
trimesh>=3.9.0
requests>=2.28.0

# New requirements for Hugging Face Transformers
transformers>=4.30.0
accelerate>=0.20.0
bitsandbytes>=0.40.0  # For quantization support